refactor: moved K8Client to k8_client folder
