+++ src/core/kube/k8_client/k8_client.ts
@@ -2,42 +2,41 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import * as k8s from '@kubernetes/client-node';
-import {type V1Lease} from '@kubernetes/client-node';
-import {Flags as flags} from '../../commands/flags.js';
-import {MissingArgumentError, SoloError} from './../errors.js';
-import {ConfigManager} from './../config_manager.js';
-import {SoloLogger} from './../logging.js';
-import {type TarCreateFilter} from '../../types/aliases.js';
+import {Flags as flags} from '../../../commands/flags.js';
+import {MissingArgumentError, SoloError} from '../../errors.js';
+import {ConfigManager} from '../../config_manager.js';
+import {SoloLogger} from '../../logging.js';
+import {type TarCreateFilter} from '../../../types/aliases.js';
 import {inject, injectable} from 'tsyringe-neo';
-import {patchInject} from './../container_helper.js';
-import {type K8} from './k8.js';
-import {type Namespaces} from './resources/namespace/namespaces.js';
-import {type NamespaceName} from './resources/namespace/namespace_name.js';
-import {K8ClientClusters} from './k8_client/k8_client_clusters.js';
-import {type Clusters} from './clusters.js';
-import {type ConfigMaps} from './config_maps.js';
-import {K8ClientConfigMaps} from './k8_client/k8_client_config_maps.js';
-import {type ContainerRef} from './container_ref.js';
-import {K8ClientContainers} from './k8_client/k8_client_containers.js';
-import {type Containers} from './containers.js';
-import {type Contexts} from './contexts.js';
-import {K8ClientContexts} from './k8_client/k8_client_contexts.js';
-import {K8ClientPods} from './k8_client/resources/pod/k8_client_pods.js';
-import {type Pods} from './resources/pod/pods.js';
-import {K8ClientBase} from './k8_client/k8_client_base.js';
-import {type Services} from './services.js';
-import {K8ClientServices} from './k8_client/k8_client_services.js';
-import {type Pvcs} from './resources/pvc/pvcs.js';
-import {K8ClientPvcs} from './k8_client/resources/pvc/k8_client_pvcs.js';
-import {type Leases} from './resources/lease/leases.js';
-import {K8ClientLeases} from './k8_client/resources/lease/k8_client_leases.js';
-import {K8ClientNamespaces} from './k8_client/resources/namespace/k8_client_namespaces.js';
-import {K8ClientIngressClasses} from './k8_client/k8_client_ingress_classes.js';
-import {type IngressClasses} from './ingress_classes.js';
-import {type Secrets} from './secrets.js';
-import {K8ClientSecrets} from './k8_client/k8_client_secrets.js';
-import {PvcRef} from './resources/pvc/pvc_ref.js';
-import {PvcName} from './resources/pvc/pvc_name.js';
+import {patchInject} from '../../container_helper.js';
+import {type K8} from '../k8.js';
+import {type Namespaces} from '../resources/namespace/namespaces.js';
+import {type NamespaceName} from '../resources/namespace/namespace_name.js';
+import {K8ClientClusters} from '../k8_client/k8_client_clusters.js';
+import {type Clusters} from '../clusters.js';
+import {type ConfigMaps} from '../config_maps.js';
+import {K8ClientConfigMaps} from '../k8_client/k8_client_config_maps.js';
+import {type ContainerRef} from '../container_ref.js';
+import {K8ClientContainers} from '../k8_client/k8_client_containers.js';
+import {type Containers} from '../containers.js';
+import {type Contexts} from '../contexts.js';
+import {K8ClientContexts} from '../k8_client/k8_client_contexts.js';
+import {K8ClientPods} from '../k8_client/resources/pod/k8_client_pods.js';
+import {type Pods} from '../resources/pod/pods.js';
+import {K8ClientBase} from '../k8_client/k8_client_base.js';
+import {type Services} from '../services.js';
+import {K8ClientServices} from '../k8_client/k8_client_services.js';
+import {type Pvcs} from '../resources/pvc/pvcs.js';
+import {K8ClientPvcs} from '../k8_client/resources/pvc/k8_client_pvcs.js';
+import {type Leases} from '../resources/lease/leases.js';
+import {K8ClientLeases} from '../k8_client/resources/lease/k8_client_leases.js';
+import {K8ClientNamespaces} from '../k8_client/resources/namespace/k8_client_namespaces.js';
+import {K8ClientIngressClasses} from '../k8_client/k8_client_ingress_classes.js';
+import {type IngressClasses} from '../ingress_classes.js';
+import {type Secrets} from '../secrets.js';
+import {K8ClientSecrets} from '../k8_client/k8_client_secrets.js';
+import {PvcRef} from '../resources/pvc/pvc_ref.js';
+import {PvcName} from '../resources/pvc/pvc_name.js';
 
 /**
  * A kubernetes API wrapper class providing custom functionalities required by solo
