+++ test/e2e/integration/commands/init.test.ts
@@ -10,7 +10,7 @@ import {Helm} from '../../../../src/core/helm.js';
 import {ChartManager} from '../../../../src/core/chart_manager.js';
 import {ConfigManager} from '../../../../src/core/config_manager.js';
 import {type K8} from '../../../../src/core/kube/k8.js';
-import {K8Client} from '../../../../src/core/kube/k8_client.js';
+import {K8Client} from '../../../../src/core/kube/k8_client/k8_client.js';
 import {LocalConfig} from '../../../../src/core/config/local_config.js';
 import {KeyManager} from '../../../../src/core/key_manager.js';
 import {LeaseManager} from '../../../../src/core/lease/lease_manager.js';
