+++ test/e2e/integration/core/k8_e2e.test.ts
@@ -20,7 +20,7 @@ import crypto from 'crypto';
 import {PodName} from '../../../../src/core/kube/resources/pod/pod_name.js';
 import {Duration} from '../../../../src/core/time/duration.js';
 import {container} from 'tsyringe-neo';
-import {type K8Client} from '../../../../src/core/kube/k8_client.js';
+import {type K8Client} from '../../../../src/core/kube/k8_client/k8_client.js';
 import {NamespaceName} from '../../../../src/core/kube/resources/namespace/namespace_name.js';
 import {PodRef} from '../../../../src/core/kube/resources/pod/pod_ref.js';
 import {ContainerName} from '../../../../src/core/kube/container_name.js';
